pyproject.toml
[tool.poetry]
name = "lacreiprocessoseletivo"
version = "0.1.0"
description = ""
authors = ["walefy <walefyd1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.1"
django-filter = "^24.2"
markdown = "^3.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
